+++ cli/cli/commands/enclave/inspect/modules.go
@@ -2,71 +2,83 @@ package inspect
 
 import (
 	"context"
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_manager"
+	"fmt"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/container_status"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/module"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/output_printers"
-	"github.com/kurtosis-tech/object-attributes-schema-lib/forever_constants"
-	"github.com/kurtosis-tech/object-attributes-schema-lib/schema"
 	"github.com/kurtosis-tech/stacktrace"
+	"sort"
+	"strings"
 )
 
 const (
-	moduleGUIDColHeader                                    = "GUID"
-	moduleHostMachinePortBindingsColHeader                 = "LocalPortBindings"
-	shouldShowStoppedContainersWhenGettingModuleContainers = true
-
-	noModuleHostPortBindingsPlaceholder = "<none>"
+	moduleGUIDColHeader  = "GUID"
+	modulePortsColHeader = "Ports"
 )
 
-func printModules(ctx context.Context, dockerManager *docker_manager.DockerManager, enclaveId string) error {
-	moduleLabels := getLabelsForListEnclaveModules(enclaveId)
+func printModules(ctx context.Context, kurtosisBackend backend_interface.KurtosisBackend, enclaveId enclave.EnclaveID) error {
 
-	containers, err := dockerManager.GetContainersByLabels(ctx, moduleLabels, shouldShowStoppedContainersWhenGettingModuleContainers)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred getting module containers by labels: '%+v'", moduleLabels)
+	moduleFilters := &module.ModuleFilters{
+		EnclaveIDs: map[enclave.EnclaveID]bool{
+			enclaveId: true,
+		},
+		Statuses: map[container_status.ContainerStatus]bool{
+			container_status.ContainerStatus_Stopped: true,
+			container_status.ContainerStatus_Running: true,
+		},
 	}
 
-	tablePrinter := output_printers.NewTablePrinter(moduleGUIDColHeader, moduleHostMachinePortBindingsColHeader)
-	sortedContainers, err := sortContainersByGUID(containers)
+	modules, err := kurtosisBackend.GetModules(ctx, moduleFilters)
 	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred sorting module containers by GUID")
+		return stacktrace.Propagate(err, "An error occurred getting modules using filters '%+v'", moduleFilters)
 	}
-	for _, container := range sortedContainers {
-		containerGuid, found := container.GetLabels()[schema.GUIDLabel]
-		if !found {
-			return stacktrace.NewError("No '%v' container label was found in container ID '%v' with labels '%+v'", schema.GUIDLabel, container.GetId(), container.GetLabels())
-		}
-		hostPortBindingsStrings := getContainerHostPortBindingStrings(container)
 
-		firstHostPortBindingStr := noModuleHostPortBindingsPlaceholder
-		if hostPortBindingsStrings != nil {
-			firstHostPortBindingStr = hostPortBindingsStrings[0]
-			hostPortBindingsStrings = hostPortBindingsStrings[1:]
-		}
-		if err := tablePrinter.AddRow(containerGuid, firstHostPortBindingStr); err != nil {
-			return stacktrace.NewError(
-				"An error occurred adding row for module container '%v' to the table printer",
-				containerGuid,
+	tablePrinter := output_printers.NewTablePrinter(moduleGUIDColHeader, modulePortsColHeader)
+	sortedModules := getSortedModuleSliceFromModulesMap(modules)
+
+	for _, module := range sortedModules {
+
+		portString := fmt.Sprintf(
+			"%v/%v",
+			 module.GetPrivatePort().GetNumber(),
+			 strings.ToLower(module.GetPrivatePort().GetProtocol().String()),
+		)
+		if module.GetStatus() == container_status.ContainerStatus_Running {
+			portString = portString + fmt.Sprintf(
+				" -> %v:%v",
+				module.GetPublicIp(),
+				module.GetPublicPort().GetNumber(),
 			)
 		}
 
-		for _, additionalHostPortBindingStr := range hostPortBindingsStrings {
-			if err := tablePrinter.AddRow("", additionalHostPortBindingStr); err != nil {
-				return stacktrace.NewError(
-					"An error occurred adding additional host port binding '%v' row for module container '%v' to the table printer",
-					additionalHostPortBindingStr,
-					containerGuid,
-				)
-			}
+		moduleGuidStr := string(module.GetGUID())
+
+		if err := tablePrinter.AddRow(moduleGuidStr, portString); err != nil {
+			return stacktrace.NewError(
+				"An error occurred adding row for module GUID '%v' to the table printer",
+				moduleGuidStr,
+			)
 		}
+
 	}
+
 	tablePrinter.Print()
 
 	return nil
 }
 
-func getLabelsForListEnclaveModules(enclaveId string) map[string]string {
-	labels := map[string]string{}
-	labels[forever_constants.ContainerTypeLabel] = schema.ContainerTypeModuleContainer
-	labels[schema.EnclaveIDContainerLabel] = enclaveId
-	return labels
+func getSortedModuleSliceFromModulesMap(modules map[module.ModuleGUID]*module.Module) []*module.Module {
+
+	modulesResult := make([]*module.Module, 0, len(modules))
+	for _, module := range modules {
+		modulesResult = append(modulesResult, module)
+	}
+
+	sort.Slice(modulesResult, func(i, j int) bool {
+		return modulesResult[i].GetGUID() < modulesResult[j].GetGUID()
+	})
+
+	return modulesResult
 }