+++ cli/cli/commands/service/shell/shell.go
@@ -6,16 +6,15 @@
 package shell
 
 import (
+	"bufio"
 	"context"
 	"fmt"
-	"github.com/docker/docker/api/types"
-	"github.com/docker/docker/client"
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_manager"
+	"github.com/kurtosis-tech/container-engine-lib/lib"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/command_str_consts"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/logrus_log_levels"
-	labels_helper "github.com/kurtosis-tech/kurtosis-cli/cli/helpers/service_container_labels_by_enclave_id"
 	"github.com/kurtosis-tech/kurtosis-cli/commons/positional_arg_parser"
-	"github.com/kurtosis-tech/object-attributes-schema-lib/schema"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
@@ -29,8 +28,6 @@ const (
 	kurtosisLogLevelArg                    = "kurtosis-log-level"
 	enclaveIDArg                           = "enclave-id"
 	guidArg                                = "guid"
-	shouldShowStoppedUserServiceContainers = true
-
 )
 
 var defaultKurtosisLogLevel = logrus.InfoLevel.String()
@@ -76,78 +73,34 @@ func run(cmd *cobra.Command, args []string) error {
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred parsing the positional args")
 	}
-	enclaveID := parsedPositionalArgs[enclaveIDArg]
-	guid := parsedPositionalArgs[guidArg]
+	enclaveIdStr := parsedPositionalArgs[enclaveIDArg]
+	enclaveId := enclave.EnclaveID(enclaveIdStr)
+	guidStr := parsedPositionalArgs[guidArg]
+	guid := service.ServiceGUID(guidStr)
 
-	// TODO Remove once KurtosisBackend can create an interactive shell on a container
-	dockerClient, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
+	kurtosisBackend, err := lib.GetLocalDockerKurtosisBackend()
 	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred creating the Docker client")
+		return stacktrace.Propagate(err, "An error occurred getting local Docker Kurtosis backend")
 	}
-	dockerManager := docker_manager.NewDockerManager(
-		dockerClient,
-	)
 
-	labels := labels_helper.GetUserServiceContainerLabelsWithEnclaveID(enclaveID)
-	labels[schema.GUIDLabel] = guid
-
-	containers, err := dockerManager.GetContainersByLabels(ctx, labels, shouldShowStoppedUserServiceContainers)
+	conn, err := kurtosisBackend.GetConnectionWithUserService(ctx, enclaveId, guid)
 	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred getting containers by labels: '%+v'", labels)
+		return stacktrace.Propagate(err, "An error occurred getting connection with user service with GUID '%v' in enclave '%v'", guid, enclaveId)
 	}
+	defer conn.Close()
 
-	if containers == nil || len(containers) == 0 {
-		logrus.Errorf("No service containers found for enclave with ID '%v'", enclaveID)
-		return stacktrace.NewError("No service containers found for service with GUID '%v' in enclave '%v'", guid, enclaveID)
-	}
-
-	if len(containers) > 1 {
-		return stacktrace.NewError("Only one container with enclave ID '%v' and GUID '%v' should exist but found '%v' containers with these properties", enclaveID, guid, len(containers))
-	}
-
-	serviceContainer := containers[0]
-
-	config := types.ExecConfig{
-		AttachStdin:  true,
-		Tty:          true,
-		AttachStderr: true,
-		AttachStdout: true,
-		Detach:       false,
-		Cmd:          commandToRun,
-	}
-
-	response, err := dockerClient.ContainerExecCreate(ctx, serviceContainer.GetId(), config)
-
-	if err != nil {
-		return stacktrace.Propagate(err, "an error occurred while creating the ContainerExec")
-	}
-
-	execID := response.ID
-	if execID == "" {
-		return stacktrace.NewError("the Exec ID was empty")
-	}
-
-	execStartCheck := types.ExecStartCheck{
-		Detach: false,
-		Tty:    true,
-	}
-
-	hijackedResponse, err := dockerClient.ContainerExecAttach(ctx, execID, execStartCheck)
-	if err != nil {
-		return stacktrace.Propagate(err, "There was an error while attaching to the ContainerExec")
-	}
-	defer hijackedResponse.Close()
+	newReader := bufio.NewReader(conn)
 
 	// From this point on down, I don't know why it works.... but it does
 	// I just followed the solution here: https://stackoverflow.com/questions/58732588/accept-user-input-os-stdin-to-container-using-golang-docker-sdk-interactive-co
 	// This channel is being used to know the user exited the ContainerExec
 	finishChan := make(chan bool)
 	go func() {
-		io.Copy(os.Stdout, hijackedResponse.Reader)
+		io.Copy(os.Stdout, newReader)
 		finishChan <- true
 	}()
-	go io.Copy(os.Stderr, hijackedResponse.Reader)
-	go io.Copy(hijackedResponse.Conn, os.Stdin)
+	go io.Copy(os.Stderr, newReader)
+	go io.Copy(conn, os.Stdin)
 
 	stdinFd := int(os.Stdin.Fd())
 	var oldState *terminal.State